Move shared enums to auto_tags
Also removed some unused imports.
