+++ test/configuration_v1/test_auto_tags.py
@@ -12,7 +12,7 @@ from dynatrace.configuration_v1.auto_tags import (
 )
 from dynatrace.environment_v2.schemas import ConfigurationMetadata
 from dynatrace.pagination import PaginatedList
-from dynatrace.configuration_v1.dt_enums import ConditionKeyAttribute, RuleType
+from dynatrace.configuration_v1.auto_tags import ConditionKeyAttribute, RuleType
 
 
 def test_list(dt: Dynatrace):
